move _parse_point_input and _parse_track_input to timeseries
